Use evil to support protobuf strings in TypeAdapterProtobuf
In NigoriMessages protobuf strings are stored in a field of type Object so that that
can use either String of ByteString as the implementation, however this causes a type
error when calling the set method. So we make a (potentially false) assumption that
all fields of type Object in NigoriMessages have that type because they actually
should have Strings set.

Signed-off-by: Daniel Thomas <drt24@cam.ac.uk>