RFC: Specify that the signature contains the command for the message
Every message sent to the server is sent as a specific command therefore
the command string should be part of the signed content to prevent the
possibility of commands being maliciously send to the wrong endpoint and having
undesired concequences.

This is what we do in the Java implementation (though we might need to alter
the exact way we do that to comply with this spec).